uhtml/reactive 🦄
This MR brings reactivity out of the box via signals based libraries.

You can bring your ogn signals based library or use the `uhtml/preactive`
export which already bundles `@preact/signals-core` in it.

The main difference with the *reactive* export, or its *preactive* one,
is that the *render* function needs the "what to render" second argument
to be a callback, otherwise signals don't get a chance to side effect.